Add/Delete using jQuery - also a more task
In previous Practice of add/remove using javascript we binded delete button to the new list items and the current were already with delete button

We did a little bit same in this using jQuery. jQuery helped alot as it saves times and lines of codes and it was easy too.
We add the data, Okay
We delete the li node without having a delete button just by picking the id of ul and calling remove() function on the particular node.

There was a problem in new li elemtens as there was missing remove binding, we handled that and did binding to new li elements.

Then we modified as we did a single binding instead of duplicate binding at two places i.e. repeatition of code.

Done.

Task:
Now focus on how to update/edit the content/element using jQuery